Aplicación/Lab2/Back/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
requests>=2.28.0
gunicorn>=20.1.0  # Para despliegue en servidores como Heroku
whitenoise>=6.0.0  # Para servir archivos estáticos
psycopg2-binary>=2.9.0  # Si usas PostgreSQL como base de datos
django-cors-headers>=3.11.0  # Para manejar CORS en APIs públicas
django-environ>=0.10.0  # Para manejar variables de entorno
pillow>=9.0.0  # Si necesitas manipulación de imágenes
